[LG-80] Move dtos to other package
