+++ order/src/main/java/ru/sinitsynme/logistapi/rest/dto/clientOrganization/ChangeClientOrganizationStatusRequestDto.java
@@ -1,4 +1,4 @@
-package ru.sinitsynme.logistapi.rest.dto;
+package ru.sinitsynme.logistapi.rest.dto.clientOrganization;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
