+++ order/src/main/java/ru/sinitsynme/logistapi/rest/dto/clientOrganization/ClientOrganizationEditRequestDto.java
@@ -1,10 +1,10 @@
-package ru.sinitsynme.logistapi.rest.dto;
+package ru.sinitsynme.logistapi.rest.dto.clientOrganization;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.sinitsynme.logistapi.entity.enums.OrganizationStatus;
+import ru.sinitsynme.logistapi.rest.dto.AddressRequestDto;
 
 import java.util.UUID;
 
@@ -12,15 +12,14 @@ import java.util.UUID;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class ClientOrganizationResponseDto {
+public class ClientOrganizationEditRequestDto {
 
     private UUID clientId;
-    private AddressResponseDto addressResponseDto;
+    private AddressRequestDto addressRequestDto;
     private String name;
-    private String inn;
     private String bik;
     private String clientAccount;
     private String bankName;
     private String correspondentAccount;
-    private OrganizationStatus organizationStatus;
+
 }
