+++ order/src/main/java/ru/sinitsynme/logistapi/rest/dto/clientOrganization/ClientOrganizationResponseDto.java
@@ -1,9 +1,11 @@
-package ru.sinitsynme.logistapi.rest.dto;
+package ru.sinitsynme.logistapi.rest.dto.clientOrganization;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import ru.sinitsynme.logistapi.entity.enums.OrganizationStatus;
+import ru.sinitsynme.logistapi.rest.dto.AddressResponseDto;
 
 import java.util.UUID;
 
@@ -11,14 +13,15 @@ import java.util.UUID;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class ClientOrganizationEditRequestDto {
+public class ClientOrganizationResponseDto {
 
     private UUID clientId;
-    private AddressRequestDto addressRequestDto;
+    private AddressResponseDto addressResponseDto;
     private String name;
+    private String inn;
     private String bik;
     private String clientAccount;
     private String bankName;
     private String correspondentAccount;
-
+    private OrganizationStatus organizationStatus;
 }
