+++ order/src/main/java/ru/sinitsynme/logistapi/service/ClientOrganizationService.java
@@ -15,8 +15,8 @@ import ru.sinitsynme.logistapi.entity.enums.OrganizationStatus;
 import ru.sinitsynme.logistapi.mapper.AddressMapper;
 import ru.sinitsynme.logistapi.mapper.ClientOrganizationMapper;
 import ru.sinitsynme.logistapi.repository.ClientOrganizationRepository;
-import ru.sinitsynme.logistapi.rest.dto.ClientOrganizationEditRequestDto;
-import ru.sinitsynme.logistapi.rest.dto.ClientOrganizationRequestDto;
+import ru.sinitsynme.logistapi.rest.dto.clientOrganization.ClientOrganizationEditRequestDto;
+import ru.sinitsynme.logistapi.rest.dto.clientOrganization.ClientOrganizationRequestDto;
 
 import java.util.UUID;
